refactor: 테스트파일 이동 및 리팩터링
